[FIXED] Replaced references to FSpGetInfo/FSpSetInfo with FSGetCatalogInfo/FSSetCatalogInfo. FSp*etInfo is deprecated under Tiger [FIXED] Reverted mac_inject to the official version (how did it ever change in the first place?!) [CHANGED] We now have some more optimisations enabled. Virtue seems to use a little less memory than it has been, but I'd say the memory usage isn't going to get much better while there are 3 frameworks, a static library and a bundle in the mix. Re-factoring (or at least simplification) will need to be put on the cards for 0.6

+++ Frameworks/Peony/src/PNWindow.m
@@ -202,13 +202,13 @@
 }
  
 - (NSString*) name {
-	static CGSValue kCGSWindowTitle = NULL; 
+	static CGSValue kCGSWindowTitle = (int)NULL; 
 
 	// we have to create the private constant 
-	if (kCGSWindowTitle == NULL) 
+	if (kCGSWindowTitle == (int)NULL) 
 		kCGSWindowTitle = CGSCreateCStringNoCopy("kCGSWindowTitle");
 
-	CGSValue oWindowTitle = NULL;
+	CGSValue oWindowTitle = (int)NULL;
 	OSStatus oResult;
 	
 	// get connection
